code_corps v0.0.1 CodeCorps.Analytics.Segment

Provides analytics tracking for Segment.com with an interface for making identify and track calls via the analytics-elixir package.

You can read more about identify and track in Segment’s documentation.

By default, in dev and test envrionments, this module will use CodeCorps.Analytics.InMemoryAPI which does not make a request to Segment’s REST API.

In prod and staging environments, the module will use CodeCorps.Analytics.SegmentAPI which will make requests to Segment’s REST API.

In your config/prod.exs you might set this like so:

config :code_corps, :analytics, CodeCorps.Analytics.SegmentAPI

Functions

get_event_name(action, model)

get_event_name(atom, struct) :: String.t

Uses the action on the record to determine the event name that should be passed in for the track call.

identify(user)

identify(CodeCorps.User.t) :: any

Calls identify in the configured API module.

track(arg, action, conn)

track({:ok, Ecto.Schema.t} | {:error, Ecto.Changeset.t}, atom, Plug.Conn.t) :: any

Calls track in the configured API module.

Receives either an :ok or :error tuple from an attempted Ecto.Repo operation.

track_sign_in(conn)

track_sign_in(Plug.Conn.t) :: any

Calls track with the “Signed In” event in the configured API module.