code_corps v0.0.1 CodeCorps.AuthToken

Represents one of the user’s many possible authentication tokens, created using Phoenix.Token.sign/4.

Many can coexist and be valid at the same time. They can be used for password resets or passwordless logins.

These tokens do expire based on the max_age value passed to Phoenix.Token.verify/4.

Functions

changeset(struct, user)

Builds a changeset based on the struct