code_corps v0.0.1 CodeCorps.DonationGoalsManager

Handles CRUD operations for donation goals.

When operations happen on CodeCorps.DonationGoal, we need to set the current donation goal on CodeCorps.Project.

The current donation goal should be the smallest value that is greater than the project’s current total donations, or falls back to the largest donation goal.

Functions

create(attributes)

set_current_goal_for_project(project)

update(donation_goal, attributes)