code_corps v0.0.1 CodeCorps.GitHub

Types

api_error_struct()

api_error_struct ::
  CodeCorps.GitHub.APIError.t |
  CodeCorps.GitHub.HTTPClientError.t |
  Poison.DecodeError.t

body()

body() :: {:multipart, list} | map

headers()

headers() :: %{optional(String.t) => String.t} | %{}

method()

method() :: :get | :post | :put | :delete | :patch

Functions

authenticated_integration_request(body, method, endpoint, headers, opts)

authenticated_integration_request(body, method, String.t, headers, list) ::
  {:ok, map} |
  {:error, api_error_struct}

A low level utility function to make an authenticated request to the GitHub API on behalf of a GitHub App or integration

request(body, method, endpoint, headers, opts)

request(body, method, String.t, headers, list) ::
  {:ok, map} |
  {:error, api_error_struct}

A low level utility function to make a direct request to the GitHub API.

user_access_token_request(code, state)

user_access_token_request(String.t, String.t) ::
  {:ok, map} |
  {:error, api_error_struct}

A low level utility function to fetch a GitHub user’s OAuth access token