code_corps v0.0.1 CodeCorps.GitHub.API

Functions

gateway()

get_all(url, headers, options)

get_all(String.t(), raw_headers(), raw_options()) ::
  {:ok, [map()]} |
  {:error, CodeCorps.GitHub.API.Errors.PaginationError.t()} |
  {:error, CodeCorps.GitHub.api_error_struct()}

opts_for(installation)

opts_for(CodeCorps.GithubAppInstallation.t()) :: list()

Get access token headers for a given CodeCorps.GithubAppInstallation.

This should only be used in instances where the full permissions of the application are needed and there is no need for attribution to a user.

opts_for(user, installation)

opts_for(CodeCorps.User.t(), CodeCorps.GithubAppInstallation.t()) :: list()

Get access token headers for a given CodeCorps.User and CodeCorps.GithubAppInstallation.

If the user does not have a github_auth_token (meaning they are not connected to GitHub), then we default to the installation which will post on behalf of the user as a bot.

request(method, url, body, headers, options)

request(CodeCorps.GitHub.method(), String.t(), raw_body(), raw_headers(), raw_options()) :: CodeCorps.GitHub.response()