code_corps v0.0.1 CodeCorps.GitHub.API.Repository

Functions for retrieving a GitHub repository’s issues, pull requests, and comments from the GitHub API.

Functions

issue_comments(github_repo)

issue_comments(CodeCorps.GithubRepo.t()) ::
  {:ok, [map()]} |
  {:error, CodeCorps.GitHub.paginated_endpoint_error()}

Retrieves comments from all issues in a github repository.

issues(github_repo)

issues(CodeCorps.GithubRepo.t()) ::
  {:ok, [map()]} |
  {:error, CodeCorps.GitHub.paginated_endpoint_error()}

Retrieves issues for a repository

All pages of records are retrieved. Closed issues are included.

pulls(github_repo)

pulls(CodeCorps.GithubRepo.t()) ::
  {:ok, [map()]} |
  {:error, CodeCorps.GitHub.paginated_endpoint_error()}

Retrieves pull requests for a repository.

All pages of records are retrieved.