code_corps v0.0.1 CodeCorps.GitHub.Adapters.Comment

Used to convert between GitHub data which represents a GitHub Issue Comment and CodeCorps.Comment as well as CodeCorps.GithubComment attributes.

Functions

to_api(comment)

to_api(CodeCorps.Comment.t()) :: map()

Converts a CodeCorps.Comment into a set of attributes suitable for creating or updating an GitHub Issue Comment through the GitHub API.

to_comment(github_comment)

to_comment(CodeCorps.GithubComment.t()) :: map()

Converts a CodeCorps.GithubComment into a set of attributes suitable for creating or updating a CodeCorps.Comment

to_github_comment(payload)

to_github_comment(map()) :: map()

Converts a GitHub Issue Comment payload into a set of attributes suitable for creating or updating a CodeCorps.GithubComment