code_corps v0.0.1 CodeCorps.GitHub.Adapters.Issue

Used to adapt a GitHub Issue payload into attributes for creating or updating a CodeCorps.GithubIssue and vice-versa.

Functions

to_api(github_issue)

to_api(CodeCorps.GithubIssue.t() | CodeCorps.Task.t()) :: map()

Converts a GithubIssue or Task into a set of attributes used to create or update an associated GitHub Issue on the GitHub API.

to_issue(payload)

to_issue(map()) :: map()

Converts a GitHub Issue payload into a set of attributes used to create or update a GithubIssue record.

to_task(github_issue)

to_task(CodeCorps.GithubIssue.t()) :: map()

Converts a GitHub Issue payload into a set of attributes used to create or update a Task record.