code_corps v0.0.1 CodeCorps.GitHub.Adapters.User

Used to adapt a GitHub issue payload into attributes for creating or updating a CodeCorps.Task.

Functions

from_github_user(payload)

from_github_user(map) :: map

Converts a Github user payload into a map of attributes suitable for creating or updating a CodeCorps.User

Any nil values are removed here. For example, we don’t want to delete an existing email just because the GitHub payload is missing that data.

The type gets transformed to match our expected values for user type.