code_corps v0.0.1 CodeCorps.GitHub.Event

In charge of marking GithubEvent records as “processing”, “processed” or “errored”, based on the outcome of processing a webhook event payload.

Types

error()

error() :: atom | Ecto.Changeset.t

Functions

start_processing(event)

start_processing(CodeCorps.GithubEvent.t) :: {:ok, CodeCorps.GithubEvent.t}

Sets record status to “processing”, marking it as being processed at this moment. Our webhook handling should skip processing payloads for events which are already being processed.

stop_processing(arg, event)

stop_processing(tuple, CodeCorps.GithubEvent.t) :: {:ok, CodeCorps.GithubEvent.t}

Sets record status to “processed” or “errored” based on the first element of first argument, which is the result tuple. The first element of the result tuple should always be either :ok, or :error. Any number of elements in the tuple is suported.