code_corps v0.0.1 CodeCorps.GitHub.Event.Installation

In charge of dealing with “Installation” GitHub Webhook events

Functions

handle(event, payload)

handle(CodeCorps.GithubEvent.t, map) :: {:ok, CodeCorps.GithubEvent.t}

Handles an “Installation” GitHub Webhook event. The event could be of subtype “created” or “deleted”. Only the “created” variant is handled at the moment.

InstallationCrreated::added will first try to find the User and the GithubAppInstallation, using information from the payload.

if neither are found, it will

create a fresh GithubAppInstallation, with status “unmatched_user”

if only the User is found, it will

create GithubAppInstallation with status “initiated_on_github”, associated with the User

if both are found, it will

fetch repositories for the installation, create GithubRepo records

mark installation as processed

find GithubAppInstallation

marks event as errored if not found

marks event as errored if payload does not have keys it needs, meaning there’s something wrong with our code and we need to updated

find or create GithubRepo records affected

At the end of the process, the associated event will me marked as “processed”. If any failure occurs, the event will be marked as “errored”.

Potential points of failure:

the installation is found, but the user is not

problem communicating with the API

the payload received is not of the expected structure

the action type is not one of the supported types (“created”)