code_corps v0.0.1 CodeCorps.GitHub.Event.Installation.MatchedUser

In charge of handling the matched user case of an installation event

Functions

handle(user, installation_attrs)

handle(map, CodeCorps.User.t) :: outcome

Handles the installation event in the case of a matched user.

This is done by attempting to find a GithubAppInstallation record belonging to the matched user, with a blank github_id field.

If no records are found, this means the installation was done from GitHub directly and it just so happens we have a connected GitHub user on record who performed the installation.

If a record is found, then we update it’s github_id field and then fetch repositories for that installation and store them locally.

The case of multiple records being found should not be possible and results in an error tuple being returned.