code_corps v0.0.1 CodeCorps.GitHub.Event.Installation.UnmatchedUser

In charge of handling the unmatched user case of an installation event

Functions

handle(payload)

handle(map) :: outcome

Handles the installation event in the case of an unmatched user.

This is done by attempting to find a GithubAppInstallation record by it’s github_id field.

If a record is not found, which is the expected case, one is created and it’s sender_github_id field is set, to be matched with a user which could later connect with github.

If a record is found, the something must have went wrong at some point, since it should not happen, but the process continues gracefuly and simply updates the record (making no actual changes in most cases).