code_corps v0.0.1 CodeCorps.GitHub.Event.InstallationRepositories

In charge of handling a GitHub Webhook payload for the InstallationRepositories event type.

https://developer.github.com/v3/activity/events/types/#installationrepositoriesevent

Functions

handle(payload)

handle(map()) ::
  CodeCorps.GitHub.Sync.installation_repositories_event_outcome() |
  {:error, :unexpected_payload}

Handles an “InstallationRepositories” GitHub Webhook event. The event could be of subtype “added” or “removed” and is handled differently based on that.

the process of handling the “added” subtype is as follows

try to match with CodeCorps.GithubAppInstallation record

sync affected CodeCorps.GithubRepo records (update, create)

the process of handling the “removed” subtype is as follows

try to match with a CodeCorps.GithubAppInstallation record

delete affected CodeCorps.GithubRepo records, respecting the rules

if the GitHub payload for a repo is not matched with a record in our database, just skip deleting it