code_corps v0.0.1 CodeCorps.GitHub.Event.IssueComment.UserLinker

In charge of finding a user to link with a Comment when processing an IssueComment webhook.

Functions

find_or_create_disassociated_user(attrs)

find_or_create_disassociated_user(map) :: {:ok, CodeCorps.User.t}

find_or_create_user(attrs)

find_or_create_user(map) :: {:ok, CodeCorps.User.t}

Finds or creates a user using information contained in a GitHub IssueComment webhook payload.

The process is as follows:

Find all affected comments and extract their user data.

Search for the user in our database.

If we match a single user, then the comment should be for that user.

If there are no matching users, then the comment was created on Github by someone who does not have a matching GitHub-connected Code Corps account. We create a placeholder user account until that GitHub user is claimed by a Code Corps user. created.

If there are multiple matching users, this is an unexpected scenario and should error out.