code_corps v0.0.1 CodeCorps.GitHub.Event.Issues

In charge of dealing with “Issues” GitHub Webhook events

https://developer.github.com/v3/activity/events/types/#issuesevent

Functions

handle(github_event, payload)

handle(CodeCorps.GithubEvent.t, map) :: outcome

Handles the “Issues” GitHub webhook

The process is as follows

validate the payload is structured as expected

try and find the appropriate GithubRepo record.

for each ProjectGithubRepo belonging to that Project

find or initialize a new Task

try and find a User, associate Task with user

commit the change as an insert or update action

Depending on the success of the process, the function will return one of

{:ok, list_of_tasks}

{:error, :not_fully_implemented} - while we’re aware of this action, we have not implemented support for it yet

{:error, :unexpected_payload} - the payload was not as expected

{:error, :unexpected_action} - the action was not of type we are aware of

{:error, :unmatched_repository} - the repository for this issue was not found

Note that it is also possible to have a matched GithubRepo, but with that record not having any ProjectGithubRepo children. The outcome of that case should NOT be an errored event, since it simply means that the GithubRepo was not linked to a Project by the Project owner. This is allowed and relatively common.