code_corps v0.0.1 CodeCorps.GitHub.SuccessAPI

A mocked github API layer which returns a default successful response for all GitHub API requests.

All tests in the test environment use this module as a mock for GitHub API requests by default.

If certain tests explicitly depend on the data returned by GitHub, they can be mocked individually using the CodeCorps.GitHub.TestHelpers.with_mock_api macro.

As support for new GitHub endpoints is added, defaults for those endpoints should be added here.

To assert a request has been made to GitHub as a result as an action, the assert_received test helper can be used:

assert_received({:get, "https://api.github.com/user", headers, body, options})

Functions

request(method, url, headers, body, options)