code_corps v0.0.1 CodeCorps.GitHub.Sync

Types

outcome()

outcome() ::
  {:ok, [CodeCorps.Comment.t()]} |
  {:ok, CodeCorps.GithubPullRequest.t()} |
  {:ok, [CodeCorps.Task.t()]} |
  {:error, :repo_not_found} |
  {:error, :fetching_issue} |
  {:error, :fetching_pull_request} |
  {:error, :multiple_issue_users_match} |
  {:error, :multiple_comment_users_match} |
  {:error, :validating_github_pull_request} |
  {:error, :validating_github_issue} |
  {:error, :validating_github_comment} |
  {:error, :validating_user} |
  {:error, :validating_tasks} |
  {:error, :validating_comments} |
  {:error, :unexpected_transaction_outcome}

Functions

issue_comment_event(payload)

Syncs a GitHub IssueComment event.

For the deleted action

Deletes the related comment records with GitHub.Sync.Comment

For any other action

Finds the CodeCorps.GithubRepo

If it’s a pull request, it fetches the pull request from the GitHub API and syncs it with GitHub.Sync.PullRequest

Syncs the issue portion of the event with GitHub.Sync.Issue

Syncs the comment portion of the event with GitHub.Sync.Comment

https://developer.github.com/v3/activity/events/types/#issuecommentevent

issue_event(payload)

Syncs a GitHub Issues event.

Finds the CodeCorps.GithubRepo

Syncs the issue portion of the event with GitHub.Sync.Issue

https://developer.github.com/v3/activity/events/types/#issuesevent

pull_request_event(payload)

Syncs a GitHub PullRequest event.

Finds the CodeCorps.GithubRepo

Fetches the issue from the GitHub API

Syncs the pull request portion of the event with GitHub.Sync.PullRequest

Syncs the issue portion of the event with GitHub.Sync.Issue, using the changes passed in from the issue fetching and the pull request syncing

https://developer.github.com/v3/activity/events/types/#pullrequestevent

sync_issue(issue, repo)

sync_issues(repo)