code_corps v0.0.1 CodeCorps.GitHub.Sync.Comment

Types

outcome()

outcome ::
  {:ok, [CodeCorps.Comment.t]} |
  {:error, :repository_not_found} |
  {:error, :validation_error_on_inserting_issue_for_task} |
  {:error, :validation_error_on_inserting_github_comment} |
  {:error, :validation_error_on_inserting_user_for_task} |
  {:error, :multiple_github_users_matched_same_cc_user_for_task} |
  {:error, :validation_error_on_inserting_user_for_comment} |
  {:error, :multiple_github_users_matched_same_cc_user_for_comment} |
  {:error, :validation_error_on_syncing_tasks} |
  {:error, :validation_error_on_syncing_comments} |
  {:error, :unexpected_transaction_outcome}

Functions

sync(payload)

sync(map) :: outcome

Syncs a GitHub comment API payload with our data.

The process is as follows:

match payload with CodeCorps.GithubRepo record using CodeCorps.GitHub.Sync.Utils.RepoFinder

match issue part of the payload with CodeCorps.User using CodeCorps.GitHub.Sync.User.RecordLinker

match comment part of the payload with a CodeCorps.User using CodeCorps.GitHub.Sync.User.RecordLinker

for each CodeCorps.ProjectGithubRepo belonging to the matched repo:

create or update CodeCorps.Task for the CodeCorps.Project

create or update CodeCorps.Comment for the CodeCorps.Task

If the sync succeeds, it will return an :ok tuple with a list of created or updated comments.

If the sync fails, it will return an :error tuple, where the second element is the atom indicating a reason.