code_corps v0.0.1 CodeCorps.GitHub.Sync.Comment.Comment

In charge of syncing CodeCorps.Comment records with a GitHub comment payload.

A single GitHub comment always matches a single CodeCorps.GithubComment, but it can match multiple CodeCorps.Comment records. This module handles creating or updating all those records.

Types

outcome()

outcome() ::
  {:ok, [CodeCorps.Comment.t()]} |
  {:error, {[CodeCorps.Comment.t()], [Ecto.Changeset.t()]}}

Functions

delete_all(github_id)

delete_all(String.t()) :: {:ok, [CodeCorps.Comment.t()]}

Deletes all Comment records related to GithubComment using the GitHub ID from a GitHub API comment payload.

Returns a list of the deleted Comment records.

sync_all(tasks, github_comment, user, payload)

sync_all([CodeCorps.Task.t()], CodeCorps.GithubComment.t(), CodeCorps.User.t(), map()) :: outcome()

Creates or updates CodeCorps.Comment records for the specified list of CodeCorps.Task records.

When provided a list of CodeCorps.Task records, a CodeCorps.GithubComment, a CodeCorps.User, and a GitHub API payload , for each CodeCorps.Task record, it creates or updates a CodeCorps.Comment record, using the provided GitHub API payload, associated to the specified CodeCorps.GithubComment and CodeCorps.User

sync_project_github_repo(project_github_repo)

Creates or updates Comment records for the specified ProjectGithubRepo.

For each GithubComment record that relates to the GithubRepo for a given ProjectGithubRepo:

Find the related Task record

Create or update the Comment record

Associate the Comment record with the User that relates to the GithubUser for the GithubComment