code_corps v0.0.1 CodeCorps.GitHub.Sync.Issue

Functions

sync(payload, github_repo)

sync(map(), CodeCorps.GithubRepo.t()) :: Ecto.Multi.t()

Performs sync of a github issue payload related to a repo.

Creates or updates a CodeCorps.GithubIssue

a CodeCorps.GithubUser is created or updated as part of the process and associated to CodeCorps.GithubIssue

the record is associated to the CodeCorps.GithubRepo

an ‘unregistered’ CodeCorps.User is created if no record with matching :github_id is found

sync(payload, github_repo, github_pull_request)

sync(map(), CodeCorps.GithubRepo.t(), CodeCorps.GithubPullRequest.t()) :: Ecto.Multi.t()

Performs sync of a github issue payload related to a repo and a pull request.

Performs work identical to sync/2, with the adition of associating the resulting CodeCorps.GithubIssue with the specified CodeCorps.GithubPullRequest.