code_corps v0.0.1 CodeCorps.GitHub.Sync.Issue.GithubIssue

In charge of finding or creating a CodeCorps.GithubIssue to link with a CodeCorps.Task when processing a GitHub Issue payload.

The only entry point is create_or_update_issue/2.

Types

result()

result() ::
  {:ok, CodeCorps.GithubIssue.t()} |
  {:error, Ecto.Changeset.t()}

Functions

create_or_update_issue(payload, github_repo, github_pull_request \\ nil)

create_or_update_issue(map(), CodeCorps.GithubRepo.t(), CodeCorps.GithubPullRequest.t() | nil) :: result()

Creates or updates a CodeCorps.GithubIssue from a github issue API payload.

The created record is associated to the provided CodeCorps.GithubRepo and, optionaly, to a provided CodeCorps.GithubPullRequest.

The created record is also associated with a matched CodeCorps.GithubUser, which is created if necessary.