code_corps v0.0.1 CodeCorps.GitHub.Sync.PullRequest

Functions

sync(changes, payload)

sync(map(), map()) :: Ecto.Multi.t()

Syncs a GitHub pull request API payload with our data.

The process is as follows:

match payload with affected CodeCorps.GithubRepo record using CodeCorps.GitHub.Sync.Utils.RepoFinder

match with CodeCorps.User using CodeCorps.GitHub.Event.PullRequest.UserLinker

for each CodeCorps.ProjectGithubRepo belonging to matched repo:

create or update CodeCorps.Task for the CodeCorps.Project

If the sync succeeds, it will return an :ok tuple with a list of created or updated tasks.

If the sync fails, it will return an :error tuple, where the second element is the atom indicating a reason.