code_corps v0.0.1 CodeCorps.GitHub.TestAPI

The most basic implementation of an API module for testing. All functions here should return :ok tuples.

If we want to test the wrapper module, we can specify custom API modules during function calls.

Functions

connect(code)

connect(String.t) :: {:ok, String.t}

Receives a code string, created on the client side of the GitHub connect process.

Returns one of:

an :ok tuple indicating a successful connect process, where the second element is the OAuth token string

an :error tuple, where the second element is an error message or a struct

Callback implementation for CodeCorps.GitHub.APIContract.connect/1.