code_corps v0.0.1 CodeCorps.GitHub.User

Used to perform user actions on the github API

Functions

connect(user, code, state)

connect(CodeCorps.User.t, String.t, String.t) ::
  {:ok, CodeCorps.User.t} |
  {:error, Ecto.Changeset.t} |
  {:error, CodeCorps.GitHub.api_error_struct}

POSTs code and state to GitHub to receive an OAuth token, then associates the given user with that OAuth token.

Also associates any orphaned GithubAppInstallation records matching their sender_github_id field with the user’s github_id

Returns one of the following:

{:ok, %CodeCorps.User{}}

{:error, %Ecto.Changeset{}}

{:error, GitHub.api_error_struct}

me(access_token, opts \\ [])

me(String.t, Keyword.t) ::
  {:ok, map} |
  {:error, CodeCorps.GitHub.api_error_struct}

Requests the currently authenticated user payload from github