code_corps v0.0.1 CodeCorps.Helpers.Policy

Holds helpers for extracting record relationships and determining roles for authorization policies.

Functions

admin_or_higher?(role)

Determines if provided string is equal to one of ["admin", "owner"]

contributor_or_higher?(role)

Determines if provided string is equal to one of ["contributor", "admin", "owner"]

get_membership(arg1, user)

Retrieves the specified user’s membership record, from a CodeCorps.Project struct or an Ecto.Changeset, containing an organization_id field, or from a CodeCorps.Organization struct

Returns CodeCorps.OrganizationMembership

get_project(arg1)

Retrieves a project record, from a model struct, or an Ecto.Changeset containing a project_id field

Returns CodeCorps.Project

get_role(changeset)

Retrieves the role field, from a CodeCorps.OrganizationMembership struct or an Ecto.Changeset

Returns :string

owner?(arg1)

Determines if provided string is equal to “owner”