code_corps v0.0.1 CodeCorps.Messages.ConversationQuery

Holds helpers to query CodeCorps.Conversation records using a map of params.

Functions

active_filter(queryable, map)

active_filter(Ecto.Queryable.t(), map()) :: Ecto.Queryable.t()

Filters CodeCorps.Conversation record queries to return only those considered to be active.

Active conversations belong either:

to a CodeCorps.Message initiated by user

to a CodeCorps.Message initiated by an admin, with at least a single conversation part

project_filter(queryable, map)

project_filter(Ecto.Queryable.t(), map()) :: Ecto.Queryable.t()

Narrows down a CodeCorps.Conversation query by project_id of the parent CodeCorps.Message, if specified in a params map

status_filter(queryable, arg2)

status_filter(Ecto.Queryable.t(), map()) :: Ecto.Queryable.t()

Filters CodeCorps.Conversation record queries by their status.

user_filter(queryable, map)

user_filter(Ecto.Queryable.t(), map()) :: Ecto.Queryable.t()

Narrows down a CodeCorps.Conversation query by user_id, if specified in a params map