code_corps v0.0.1 CodeCorps.Model

A temporary module to be used by existing Model modules, before we switch to an intent based structure which Phoenix 1.3 pushes.

Functions

__using__(opts) (macro)

When used import appropriate helper modules.