code_corps v0.0.1 CodeCorps.ProjectView

Functions

__phoenix_recompile__?()

Returns true whenever the list of templates changes in the filesystem.

__resource__()

The resource name, as an atom, for this view

__templates__()

Returns the template root alongside all templates.

attributes(data, conn)

Callback implementation for c:JaSerializer.Serializer.attributes/2.

categories(struct, conn)

description(m)

description(m, c)

format(data)

format(data, conn)

format(data, conn, opts)

icon_large_url(m)

icon_large_url(m, c)

icon_thumb_url(m)

icon_thumb_url(m, c)

id(data, c)

Callback implementation for c:JaSerializer.Serializer.id/2.

inserted_at(m)

inserted_at(m, c)

links(data, conn)

Callback implementation for c:JaSerializer.Serializer.links/2.

long_description_body(m)

long_description_body(m, c)

long_description_markdown(m)

long_description_markdown(m, c)

meta(struct, conn)

Callback implementation for c:JaSerializer.Serializer.meta/2.

organization(struct, conn)

posts(struct, conn)

project_categories(struct, conn)

project_skills(struct, conn)

relationships(struct, conn)

Callback implementation for c:JaSerializer.Serializer.relationships/2.

render(template, assigns \\ %{})

Renders the given template locally.

skills(struct, conn)

slug(m)

slug(m, c)

template_not_found(template, assigns)

Specs

template_not_found(Phoenix.Template.name, map) :: no_return

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.

title(m)

title(m, c)

type()

type(data, conn)

Callback implementation for c:JaSerializer.Serializer.type/2.

updated_at(m)

updated_at(m, c)