code_corps v0.0.1 CodeCorps.Router.Helpers

Module with named helpers generated from CodeCorps.Router.

Functions

category_path(conn_or_endpoint, action)

category_path(conn_or_endpoint, action, params)

category_path(conn_or_endpoint, action, id, params)

category_url(conn_or_endpoint, action)

category_url(conn_or_endpoint, action, params)

category_url(conn_or_endpoint, action, id, params)

comment_path(conn_or_endpoint, action)

comment_path(conn_or_endpoint, action, id)

comment_path(conn_or_endpoint, action, id, params)

comment_url(conn_or_endpoint, action)

comment_url(conn_or_endpoint, action, id)

comment_url(conn_or_endpoint, action, id, params)

organization_membership_path(conn_or_endpoint, action)

organization_membership_path(conn_or_endpoint, action, params)

organization_membership_path(conn_or_endpoint, action, id, params)

organization_membership_url(conn_or_endpoint, action)

organization_membership_url(conn_or_endpoint, action, params)

organization_membership_url(conn_or_endpoint, action, id, params)

organization_path(conn_or_endpoint, action)

organization_path(conn_or_endpoint, action, params)

organization_path(conn_or_endpoint, action, id, params)

organization_url(conn_or_endpoint, action)

organization_url(conn_or_endpoint, action, params)

organization_url(conn_or_endpoint, action, id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

preview_path(conn_or_endpoint, action)

preview_path(conn_or_endpoint, action, params)

preview_url(conn_or_endpoint, action)

preview_url(conn_or_endpoint, action, params)

project_category_path(conn_or_endpoint, action)

project_category_path(conn_or_endpoint, action, params)

project_category_path(conn_or_endpoint, action, id, params)

project_category_url(conn_or_endpoint, action)

project_category_url(conn_or_endpoint, action, params)

project_category_url(conn_or_endpoint, action, id, params)

project_path(conn_or_endpoint, action)

project_path(conn_or_endpoint, action, params)

project_path(conn_or_endpoint, action, id, params)

project_path(conn_or_endpoint, action, slug, project_slug, params)

project_skill_path(conn_or_endpoint, action)

project_skill_path(conn_or_endpoint, action, params)

project_skill_path(conn_or_endpoint, action, id, params)

project_skill_url(conn_or_endpoint, action)

project_skill_url(conn_or_endpoint, action, params)

project_skill_url(conn_or_endpoint, action, id, params)

project_task_path(conn_or_endpoint, action, project_id)

project_task_path(conn_or_endpoint, action, project_id, params)

project_task_path(conn_or_endpoint, action, project_id, id, params)

project_task_url(conn_or_endpoint, action, project_id)

project_task_url(conn_or_endpoint, action, project_id, params)

project_task_url(conn_or_endpoint, action, project_id, id, params)

project_url(conn_or_endpoint, action)

project_url(conn_or_endpoint, action, params)

project_url(conn_or_endpoint, action, id, params)

project_url(conn_or_endpoint, action, slug, project_slug, params)

role_path(conn_or_endpoint, action)

role_path(conn_or_endpoint, action, params)

role_path(conn_or_endpoint, action, id, params)

role_skill_path(conn_or_endpoint, action)

role_skill_path(conn_or_endpoint, action, params)

role_skill_path(conn_or_endpoint, action, id, params)

role_skill_url(conn_or_endpoint, action)

role_skill_url(conn_or_endpoint, action, params)

role_skill_url(conn_or_endpoint, action, id, params)

role_url(conn_or_endpoint, action)

role_url(conn_or_endpoint, action, params)

role_url(conn_or_endpoint, action, id, params)

skill_path(conn_or_endpoint, action)

skill_path(conn_or_endpoint, action, params)

skill_path(conn_or_endpoint, action, id, params)

skill_url(conn_or_endpoint, action)

skill_url(conn_or_endpoint, action, params)

skill_url(conn_or_endpoint, action, id, params)

slugged_route_path(conn_or_endpoint, action, slug)

slugged_route_path(conn_or_endpoint, action, slug, params)

slugged_route_url(conn_or_endpoint, action, slug)

slugged_route_url(conn_or_endpoint, action, slug, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

stripe_auth_path(conn_or_endpoint, action, id)

stripe_auth_path(conn_or_endpoint, action, id, params)

stripe_auth_url(conn_or_endpoint, action, id)

stripe_auth_url(conn_or_endpoint, action, id, params)

task_comment_path(conn_or_endpoint, action, task_id)

task_comment_path(conn_or_endpoint, action, task_id, params)

task_comment_path(conn_or_endpoint, action, task_id, id, params)

task_comment_url(conn_or_endpoint, action, task_id)

task_comment_url(conn_or_endpoint, action, task_id, params)

task_comment_url(conn_or_endpoint, action, task_id, id, params)

task_path(conn_or_endpoint, action)

task_path(conn_or_endpoint, action, params)

task_path(conn_or_endpoint, action, id, params)

task_url(conn_or_endpoint, action)

task_url(conn_or_endpoint, action, params)

task_url(conn_or_endpoint, action, id, params)

token_path(conn_or_endpoint, action)

token_path(conn_or_endpoint, action, params)

token_url(conn_or_endpoint, action)

token_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_category_path(conn_or_endpoint, action)

user_category_path(conn_or_endpoint, action, params)

user_category_path(conn_or_endpoint, action, id, params)

user_category_url(conn_or_endpoint, action)

user_category_url(conn_or_endpoint, action, params)

user_category_url(conn_or_endpoint, action, id, params)

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, id, params)

user_role_path(conn_or_endpoint, action)

user_role_path(conn_or_endpoint, action, params)

user_role_path(conn_or_endpoint, action, id, params)

user_role_url(conn_or_endpoint, action)

user_role_url(conn_or_endpoint, action, params)

user_role_url(conn_or_endpoint, action, id, params)

user_skill_path(conn_or_endpoint, action)

user_skill_path(conn_or_endpoint, action, params)

user_skill_path(conn_or_endpoint, action, id, params)

user_skill_url(conn_or_endpoint, action)

user_skill_url(conn_or_endpoint, action, params)

user_skill_url(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, id, params)