code_corps v0.0.1 CodeCorps.Services.DonationGoalsService

Handles CRUD operations for donation goals.

When operations happen on CodeCorps.DonationGoal, we need to set the current donation goal on CodeCorps.Project.

The current donation goal should be the smallest value that is greater than the project’s current total donations, or falls back to the largest donation goal.

Functions

create(attributes)

create(map) :: tuple

Creates the CodeCorps.DonationGoal by wrapping the following in a transaction:

Inserting the donation goal

Updating the sibling goals with update_related_goals/1

update(donation_goal, attributes)

Updates the CodeCorps.DonationGoal by wrapping the following in a transaction:

Updating the donation goal

Updating the sibling goals with update_related_goals/1

update_project_goals(project)

Updates all CodeCorpes.DonationGoal records for a project.

To be used when something related to the project’s donation goals changes, but not any of the donation goals directly.

For example, a customer could update their subscription, which would change the total amount in monthly donations to a project, so the current goal might need updating.

It updates all goals for a project by

finding the current goal for a project

setting current for each of its siblings to false

setting current for the current goal to true

update_related_goals(arg1)

Updates sibling goals for a CodeCorps.DonationGoal by:

Finding this goal’s project

Finding the current goal for that project

Setting current on every sibling goal to false

Setting current on the found goal to true

The current goal is either:

The goal with the lowest amount greater than the project’s total raised;

Failing that, the highest goal less than the amount raised

For example, if the project has goals of $100 and $200, and has raised $150, then the current goal is $200.

If the project has a goal of $100 only and has raised $150, then the current goal is $100; the project is over that goal, but it is still the current one. It is therefore possible for a goal to have both current and achieved be true.