code_corps v0.0.1 CodeCorps.StripeConnectAccountView

Functions

__phoenix_recompile__?()

Returns true whenever the list of templates changes in the filesystem.

__resource__()

The resource name, as an atom, for this view

__templates__()

Returns the template root alongside all templates.

access_code(m)

access_code(m, c)

attributes(data, conn)

Callback implementation for JaSerializer.Serializer.attributes/2.

business_name(m)

business_name(m, c)

business_url(m)

business_url(m, c)

can_accept_donations(m)

can_accept_donations(m, c)

charges_enabled(m)

charges_enabled(m, c)

country(m)

country(m, c)

default_currency(m)

default_currency(m, c)

details_submitted(m)

details_submitted(m, c)

display_name(m)

display_name(m, c)

email(m)

email(m, c)

format(data)

format(data, conn)

format(data, conn, opts)

id(data, c)

Callback implementation for JaSerializer.Serializer.id/2.

id_from_stripe(m)

id_from_stripe(m, c)

inserted_at(m)

inserted_at(m, c)

links(data, conn)

Callback implementation for JaSerializer.Serializer.links/2.

managed(m)

managed(m, c)

meta(struct, conn)

Callback implementation for JaSerializer.Serializer.meta/2.

organization(struct, conn)

preload(query, conn, arg3)

Callback implementation for JaSerializer.Serializer.preload/3.

relationships(struct, conn)

Callback implementation for JaSerializer.Serializer.relationships/2.

render(template, assigns \\ %{})

Renders the given template locally.

support_email(m)

support_email(m, c)

support_phone(m)

support_phone(m, c)

support_url(m)

support_url(m, c)

template_not_found(template, assigns)

template_not_found(Phoenix.Template.name, map) :: no_return

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.

transfers_enabled(m)

transfers_enabled(m, c)

type()

type(data, conn)

Callback implementation for JaSerializer.Serializer.type/2.

updated_at(m)

updated_at(m, c)