code_corps v0.0.1 CodeCorps.StripeEvent

Represents a reference to single Stripe API Event object

Fields

endpoint - “connect” or “platform”

id_from_stripe - Stripe’s id

status - “unprocessed”, “processed”, or “errored”

Note on status

When the event is received via a webhook, it is stored as “unprocessed”. If during processing, there is an issue, it is set to “errored”. Once successfuly processed, it is set to “processed”.

There are cases where Stripe can send multiple webhooks for the same event, so when such a request is received, an event that is “errored” or “unprocessed” can be processed again, while a “processed” event is ignored.

Types

t()

t() :: %CodeCorps.StripeEvent{__meta__: term, endpoint: term, id: term, id_from_stripe: term, ignored_reason: term, inserted_at: term, object_id: term, object_type: term, status: term, type: term, updated_at: term, user_id: term}

Functions

create_changeset(struct, params \\ %{})

Builds a changeset for storing a new event reference into the database. The status field is set to “unprocessed” by default.

update_changeset(struct, params)

Builds a changeset for updating the status of an existing event reference. Accepts :status only and ensures it’s one of “unprocessed”, “processed” or “errored”.