code_corps v0.0.1 CodeCorps.StripeService.StripeConnectAccountService

Functions

create(attributes)

Used to create a remote Stripe.Account record as well as an associated local StripeConnectAccount record.

update(local_account, attributes)

Used to update both the local StripeConnectAccount as well as the remote Stripe.Account, using attributes sent by the client

update_from_stripe(id_from_stripe)

Used to update the local StripeConnectAccount record using data retrieved from the Stripe API