code_corps v0.0.1 CodeCorps.StripeService.StripePlatformCustomerService

Used to perform actions on StripePlatformCustomer records while propagating to and from associated Stripe.Customer records.

Functions

create(attributes)

update(customer, attributes)

Updates a CodeCorps.StripePlatformCustomer local and associated %Stripe.Customer{} API record with provided attributes Returns -, %Stripe.Customer{}}if everything was updated -}-if there was a validation issue while updating the local record -}- if there was a problem with updating the API record -` -if something unexpected went wrong

update_from_stripe(id_from_stripe)

Updates a CodeCorps.StripePlatformCustomer local record using %Stripe.Customer{} API record retrieved from API using the provided Stripe ID Returns -}if the local record was updated -- If there was no record witht he specified Stripe ID -}-if there was a validation issue while updating the local record -}- if there was a problem with retrieving the API record -` -if something unexpected went wrong