code_corps v0.0.1 CodeCorps.StripeService.Validators.ProjectSubscribable

Ensures a CodeCorps.Project is able to receive subscriptions.

Functions

validate(project)

Determines if the provided CodeCorps.Project is able to get a subscription by a CodeCorps.User

For a project to be able to receive subscriptions, it needs to have proper associations set up.

These are:

StripeConnectPlan

Organization with a StripeConnectAccount

If the project has these relationships set up, it returns {:ok, project}

In any other case, it returns {:error, :project_not_ready}