code_corps v0.0.1 CodeCorps.StripeService.Validators.UserCanSubscribe

Ensures a CodeCorps.User is able to subscribe to a CodeCorps.Project.

Functions

validate(user)

Determines if the provided CodeCorps.User is able to subscribe to a CodeCorps.Project

For a user to be able to create subscriptions, they need to have their associated records properly set up

These are:

StripePlatformCard

StripePlatformCustomer

If the user has these relationships set up, it returns {:ok, user}

In any other case, it returns {:error, :user_not_ready}