code_corps v0.0.1 CodeCorps.StripeService.WebhookProcessing.ConnectEventHandler

In charge of handling Stripe Connect webhooks

Functions

handle_event(attributes)

Handles Stripe Connect webhooks

Returns

The result of calling the specific handlers handle/1 function. This result ought ot be a tupple, in which the first member is :ok, followed by one or more other elements, usually modified records.

{:ok, :unhandled_event} if the specific event is not supported yet or at all