code_corps v0.0.1 CodeCorps.StripeService.WebhookProcessing.IgnoredEventHandler

Functions

handle(local_event, handler)

handle(CodeCorps.StripeEvent.t, module) :: {:ok, CodeCorps.StripeEvent.t}

Takes in a CodeCorps.StripeEvent to be processed as “ignored”. Determines the reason for ignoring the event, then updates the record to status: "ignored" and ignored_reason: inferred_message

Returns `{:ok, %CodeCorps.StripeEvent{}}

ignored_event_types(arg)

ignored_event_types(module) :: list

Returns a list of event types which are being explicitly ignored by the application.

should_handle?(type, handler)

should_handle?(String.t, module) :: boolean

Determines if an event type should be handled by MODULE

Returns true or false depending on specified type