code_corps v0.0.1 CodeCorps.Task.Query

Holds queries used to retrieve a list of, or a single Task record from the database, using a provided map of parameters/filters.

Functions

find(map)

find(map) :: Ecto.Queryable.t

Returns a Task record retrived using a set of parameters.

This set can be

a combination of project_id and number

a combination of task_list_id and number

an id

list(params)

list(map) :: [Project.t]

Returns a list of Task records, filtered by a map of parameters.

Accepted parameters are a project_id, or a list of comma separated task_list_ids, combined with a status.

The records are returned ordered by the :order field, ascending.