code_corps v0.0.1 CodeCorps.Transition.GithubAppInstallationState

Governs the transitioning from one state to the next.

The possible state values are:

initiated_on_code_corps - The user clicks a button or link in the Code Corps UI and is redirected to GitHub to install. This process creates the GithubAppInstallation record before redirecting.

initiated_on_github - We receive an installation webhook with a matching user, but an installation was never created on Code Corps. This can happen because GitHub does not assume an app will only be installed at a starting point outside GitHub.

processing - When the installation webhook is received, was matched, and is now processing.

processed - The integration process is completed. There is now a working installation attached to the project.

unmatched_user - When we receive a webhook but there is no Code Corps user matching the given GitHub user’s GitHub id. This is sent in the sender key of the installation event by GitHub.

It is possible to resolve more problematic states like unmatched_user when the user provides more information, e.g. connecting their GitHub account. The transitions below allow for some of these edge cases.

Functions

next(current_state, next_state)