code_corps v0.0.1 CodeCorps.User

This module defines a user of the Code Corps app.

Functions

apply_state_transition(changeset, map)

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

check_email_availability(email)

check_username_availability(username)

registration_changeset(struct, params)

Builds a changeset for registering the user.

update_changeset(struct, params)