code_corps v0.0.1 CodeCorps.UserView

Functions

__phoenix_recompile__?()

Returns true whenever the list of templates changes in the filesystem.

__resource__()

The resource name, as an atom, for this view

__templates__()

Returns the template root alongside all templates.

attributes(data, conn)

Callback implementation for JaSerializer.Serializer.attributes/2.

biography(m)

biography(m, c)

email(m)

email(m, c)

Returns the user email or an empty string, depending on the user being rendered is the authenticated user, or some other user.

Users can only see their own emails. Everyone else’s are private.

first_name(m)

first_name(m, c)

format(data)

format(data, conn)

format(data, conn, opts)

id(data, c)

Callback implementation for JaSerializer.Serializer.id/2.

inserted_at(m)

inserted_at(m, c)

last_name(m)

last_name(m, c)

links(data, conn)

Callback implementation for JaSerializer.Serializer.links/2.

meta(struct, conn)

Callback implementation for JaSerializer.Serializer.meta/2.

organization_memberships(struct, conn)

photo_large_url(m)

photo_large_url(m, c)

photo_thumb_url(m)

photo_thumb_url(m, c)

preload(query, conn, arg3)

Callback implementation for JaSerializer.Serializer.preload/3.

relationships(struct, conn)

Callback implementation for JaSerializer.Serializer.relationships/2.

render(template, assigns \\ %{})

Renders the given template locally.

slugged_route(struct, conn)

state(m)

state(m, c)

state_transition(m)

state_transition(m, c)

stripe_connect_subscriptions(struct, conn)

stripe_platform_card(struct, conn)

stripe_platform_customer(struct, conn)

template_not_found(template, assigns)

template_not_found(Phoenix.Template.name, map) :: no_return

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.

twitter(m)

twitter(m, c)

type()

type(data, conn)

Callback implementation for JaSerializer.Serializer.type/2.

updated_at(m)

updated_at(m, c)

user_categories(struct, conn)

user_roles(struct, conn)

user_skills(struct, conn)

username(m)

username(m, c)

website(m)

website(m, c)