code_corps v0.0.1 CodeCorpsWeb.OrganizationInviteController

Functions

filter(conn, query, arg3, id_list)

Callback executed for each filter param.

For example, if you wanted to optionally filter on an Article’s category and issue, your request url might look like:

/api/articles?filter[category]=elixir&filter[issue]=12

You would then want two callbacks:

def filter(_conn, query, "category", category) do
  where(query, category: category)
end

def filter(_conn, query, "issue", issue_id) do
  where(query, issue_id: issue_id)
end

You can also use guards to whitelist a handeful of attributes:

@filterable_attrs ~w(title category author_id issue_id)
def filter(_conn, query, attr, val) when attr in @filterable_attrs do
  where(query, [{String.to_existing_atom(attr), val}])
end

Anything not explicitly matched by your callbacks will be ignored.

Callback implementation for JaResource.Index.filter/4.

handle_create(conn, attributes)

Returns an unpersisted changeset or persisted model of the newly created object.

Default implementation returns the results of calling Model.changeset(%Model{}, attrs) where Model is the model defined by the JaResource.Model.model/0 callback.

The attributes argument is the result of the permitted_attributes function.

handle_create/2 can return an %Ecto.Changeset, an Ecto.Schema struct, a list of errors ({:error, [email: "is not valid"]} or a conn with any response/body.

Example custom implementation:

def handle_create(_conn, attributes) do
  Post.changeset(%Post{}, attributes, :create_and_publish)
end

Callback implementation for JaResource.Create.handle_create/2.

handle_delete(conn, model)

Returns an unpersisted changeset or persisted model representing the newly updated model.

Receives the conn and the record as found by record/2.

Default implementation returns the results of calling Repo.delete(record).

Example custom implementation:

def handle_delete(conn, record) do
  case conn.assigns[:user] do
    %{is_admin: true} -> super(conn, record)
    _                 -> send_resp(conn, 401, "nope")
  end
end

Callback implementation for JaResource.Delete.handle_delete/2.

handle_index(conn, params)

Returns the models to be represented by this resource.

Default implementation is the result of the JaResource.Records.records/2 callback. Usually a module or an %Ecto.Query{}.

The results of this callback are passed to the filter and sort callbacks before the query is executed.

handle_index/2 can alternatively return a conn with any response/body.

Example custom implementation:

def handle_index(conn, _params) do
  case conn.assigns[:user] do
    nil  -> App.Post
    user -> User.own_posts(user)
  end
end

In most cases JaResource.Records.records/1, filter/4, and sort/4 are the better customization hooks.

Callback implementation for JaResource.Index.handle_index/2.

handle_index_query(conn, query)

Callback executed to query repo.

By default this just calls all/2 on the repo. Can be customized for pagination, monitoring, etc. For example to paginate with Scrivener:

def handle_index_query(%{query_params: qp}, query) do
  repo().paginate(query, qp["page"] || %{})
end

Callback implementation for JaResource.Index.handle_index_query/2.

handle_invalid_create(conn, errors)

Returns a Plug.Conn in response to errors during create.

Default implementation sets the status to :unprocessable_entity and renders the error messages provided.

Callback implementation for JaResource.Create.handle_invalid_create/2.

handle_invalid_update(conn, errors)

Returns a Plug.Conn in response to errors during update.

Default implementation sets the status to :unprocessable_entity and renders the error messages provided.

Callback implementation for JaResource.Update.handle_invalid_update/2.

handle_show(conn, id)

Returns the model to be represented by this resource.

Default implementation is the result of the JaResource.Record.record/2 callback.

handle_show/2 can return nil to send a 404, a conn with any response/body, or a record to be serialized.

Example custom implementation:

def handle_show(conn, id) do
  Repo.get_by(Post, slug: id)
end

In most cases JaResource.Record.record/2 and JaResource.Records.records/1 are the better customization hooks.

Callback implementation for JaResource.Show.handle_show/2.

handle_update(conn, model, attributes)

Returns an unpersisted changeset or persisted model representing the newly updated model.

Receives the conn, the model as found by record/2, and the attributes argument from the permitted_attributes function.

Default implementation returns the results of calling Model.changeset(model, attrs).

handle_update/3 can return an %Ecto.Changeset, an Ecto.Schema struct, a list of errors ({:error, [email: "is not valid"]} or a conn with any response/body.

Example custom implementation:

def handle_update(conn, post, attributes) do
  current_user_id = conn.assigns[:current_user].id
  case post.author_id do
    ^current_user_id -> {:error, author_id: "you can only edit your own posts"}
    _                -> Post.changeset(post, attributes, :update)
  end
end

Callback implementation for JaResource.Update.handle_update/3.

model()

model() :: module

Must return the module implementing Ecto.Schema to be represented.

Example:

def model, do: MyApp.Models.Post

Defaults to the name of the controller, for example the controller MyApp.V1.PostController would serve the MyApp.Post model.

Used by the default implementations for handle_create/2, handle_update/3, and records/1.

Callback implementation for JaResource.Model.model/0.

permitted_attributes(conn, attrs, _)

Used to determine which attributes are permitted during create and update.

The attributes map (the second argument) is a “flattened” version including the values at data/attributes, data/type and any relationship values in data/relationships/[name]/data/id as name_id.

The third argument is the atom of the action being called.

Example:

defmodule MyApp.V1.PostController do
  use MyApp.Web, :controller
  use JaResource

  def permitted_attributes(conn, attrs, :create) do
    attrs
    |> Map.take(~w(title body type category_id))
    |> Map.merge("author_id", conn.assigns[:current_user])
  end

  def permitted_attributes(_conn, attrs, :update) do
    Map.take(attrs, ~w(title body type category_id))
  end
end

Callback implementation for JaResource.Attributes.permitted_attributes/3.

record(conn, id)

Used to get the subject of the current action

Many/most controllers will override this:

def record(%Plug.Conn{assigns: %{user_id: user_id}}, id) do
  model()
  |> where([p], p.author_id == ^user_id)
  |> Repo.get(id)
end

Callback implementation for JaResource.Record.record/2.

records(conn)

Used to get the base query of records.

Many/most controllers will override this:

def records(%Plug.Conn{assigns: %{user_id: user_id}}) do
  model()
  |> where([p], p.author_id == ^user_id)
end

Return value should be %Plug.Conn{} or an %Ecto.Query{}.

Callback implementation for JaResource.Records.records/1.

render_create(conn, model)

Returns a Plug.Conn in response to successful create.

Default implementation sets the status to :created and renders the view.

Callback implementation for JaResource.Create.render_create/2.

render_update(conn, model)

Returns a Plug.Conn in response to successful update.

Default implementation renders the view.

Callback implementation for JaResource.Update.render_update/2.

repo()

Defines the module use-ing Ecto.Repo to be used by the controller.

Defaults to the value set in config if present:

config :ja_resource,
   repo: MyApp.Repo

Default can be overridden per controller:

def repo, do: MyApp.SecondaryRepo

Callback implementation for JaResource.Repo.repo/0.

serialization_opts(conn, arg2, model_or_models)

Converts full list of params into serialization opts.

Typically this callback returns the list of fields and includes that were optionally requested by the stack.

It can also be used to add a meta to the payload, such as for scrivener pagination, on an index endpoint.

See http://github.com/AgilionApps/ja_serializer for option format.

Callback implementation for JaResource.Serializable.serialization_opts/3.

sort(conn, results, key, dir)

Callback executed for each value in the sort param.

Fourth argument is the direction as an atom, either :asc or :desc based upon the presence or not of a - prefix.

For example if you wanted to sort by date then title your request url might look like:

/api/articles?sort=-created,title

You would then want two callbacks:

def sort(_conn, query, "created", direction) do
  order_by(query, [{direction, :inserted_at}])
end

def sort(_conn, query, "title", direction) do
  order_by(query, [{direction, :title}])
end

Anything not explicitly matched by your callbacks will be ignored.

Callback implementation for JaResource.Index.sort/4.