code_corps v0.0.1 CodeCorpsWeb.Router.Helpers

Module with named helpers generated from CodeCorpsWeb.Router.

Functions

category_path(conn_or_endpoint, action)

category_path(conn_or_endpoint, action, params)

category_path(conn_or_endpoint, action, id, params)

category_url(conn_or_endpoint, action)

category_url(conn_or_endpoint, action, params)

category_url(conn_or_endpoint, action, id, params)

comment_path(conn_or_endpoint, action)

comment_path(conn_or_endpoint, action, params)

comment_path(conn_or_endpoint, action, id, params)

comment_url(conn_or_endpoint, action)

comment_url(conn_or_endpoint, action, params)

comment_url(conn_or_endpoint, action, id, params)

donation_goal_path(conn_or_endpoint, action)

donation_goal_path(conn_or_endpoint, action, params)

donation_goal_path(conn_or_endpoint, action, id, params)

donation_goal_url(conn_or_endpoint, action)

donation_goal_url(conn_or_endpoint, action, params)

donation_goal_url(conn_or_endpoint, action, id, params)

github_app_installation_path(conn_or_endpoint, action)

github_app_installation_path(conn_or_endpoint, action, params)

github_app_installation_path(conn_or_endpoint, action, id, params)

github_app_installation_url(conn_or_endpoint, action)

github_app_installation_url(conn_or_endpoint, action, params)

github_app_installation_url(conn_or_endpoint, action, id, params)

github_events_path(conn_or_endpoint, action)

github_events_path(conn_or_endpoint, action, params)

github_events_url(conn_or_endpoint, action)

github_events_url(conn_or_endpoint, action, params)

github_issue_path(conn_or_endpoint, action)

github_issue_path(conn_or_endpoint, action, params)

github_issue_path(conn_or_endpoint, action, id, params)

github_issue_url(conn_or_endpoint, action)

github_issue_url(conn_or_endpoint, action, params)

github_issue_url(conn_or_endpoint, action, id, params)

github_pull_request_path(conn_or_endpoint, action)

github_pull_request_path(conn_or_endpoint, action, params)

github_pull_request_path(conn_or_endpoint, action, id, params)

github_pull_request_url(conn_or_endpoint, action)

github_pull_request_url(conn_or_endpoint, action, params)

github_pull_request_url(conn_or_endpoint, action, id, params)

github_repo_path(conn_or_endpoint, action)

github_repo_path(conn_or_endpoint, action, params)

github_repo_path(conn_or_endpoint, action, id, params)

github_repo_url(conn_or_endpoint, action)

github_repo_url(conn_or_endpoint, action, params)

github_repo_url(conn_or_endpoint, action, id, params)

organization_github_app_installation_path(conn_or_endpoint, action)

organization_github_app_installation_path(conn_or_endpoint, action, params)

organization_github_app_installation_path(conn_or_endpoint, action, id, params)

organization_github_app_installation_url(conn_or_endpoint, action)

organization_github_app_installation_url(conn_or_endpoint, action, params)

organization_github_app_installation_url(conn_or_endpoint, action, id, params)

organization_invite_path(conn_or_endpoint, action)

organization_invite_path(conn_or_endpoint, action, params)

organization_invite_path(conn_or_endpoint, action, id, params)

organization_invite_url(conn_or_endpoint, action)

organization_invite_url(conn_or_endpoint, action, params)

organization_invite_url(conn_or_endpoint, action, id, params)

organization_path(conn_or_endpoint, action)

organization_path(conn_or_endpoint, action, params)

organization_path(conn_or_endpoint, action, id, params)

organization_url(conn_or_endpoint, action)

organization_url(conn_or_endpoint, action, params)

organization_url(conn_or_endpoint, action, id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

password_path(conn_or_endpoint, action)

password_path(conn_or_endpoint, action, params)

password_reset_path(conn_or_endpoint, action)

password_reset_path(conn_or_endpoint, action, params)

password_reset_url(conn_or_endpoint, action)

password_reset_url(conn_or_endpoint, action, params)

password_url(conn_or_endpoint, action)

password_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

preview_path(conn_or_endpoint, action)

preview_path(conn_or_endpoint, action, params)

preview_url(conn_or_endpoint, action)

preview_url(conn_or_endpoint, action, params)

project_category_path(conn_or_endpoint, action)

project_category_path(conn_or_endpoint, action, params)

project_category_path(conn_or_endpoint, action, id, params)

project_category_url(conn_or_endpoint, action)

project_category_url(conn_or_endpoint, action, params)

project_category_url(conn_or_endpoint, action, id, params)

project_github_repo_path(conn_or_endpoint, action)

project_github_repo_path(conn_or_endpoint, action, params)

project_github_repo_path(conn_or_endpoint, action, id, params)

project_github_repo_url(conn_or_endpoint, action)

project_github_repo_url(conn_or_endpoint, action, params)

project_github_repo_url(conn_or_endpoint, action, id, params)

project_path(conn_or_endpoint, action)

project_path(conn_or_endpoint, action, params)

project_path(conn_or_endpoint, action, id, params)

project_path(conn_or_endpoint, action, slug, project_slug, params)

project_skill_path(conn_or_endpoint, action)

project_skill_path(conn_or_endpoint, action, params)

project_skill_path(conn_or_endpoint, action, id, params)

project_skill_url(conn_or_endpoint, action)

project_skill_url(conn_or_endpoint, action, params)

project_skill_url(conn_or_endpoint, action, id, params)

project_task_list_path(conn_or_endpoint, action, project_id)

project_task_list_path(conn_or_endpoint, action, project_id, params)

project_task_list_path(conn_or_endpoint, action, project_id, id, params)

project_task_list_url(conn_or_endpoint, action, project_id)

project_task_list_url(conn_or_endpoint, action, project_id, params)

project_task_list_url(conn_or_endpoint, action, project_id, id, params)

project_task_path(conn_or_endpoint, action, project_id)

project_task_path(conn_or_endpoint, action, project_id, number)

project_task_path(conn_or_endpoint, action, project_id, number, params)

project_task_url(conn_or_endpoint, action, project_id)

project_task_url(conn_or_endpoint, action, project_id, number)

project_task_url(conn_or_endpoint, action, project_id, number, params)

project_url(conn_or_endpoint, action)

project_url(conn_or_endpoint, action, params)

project_url(conn_or_endpoint, action, id, params)

project_url(conn_or_endpoint, action, slug, project_slug, params)

project_user_path(conn_or_endpoint, action)

project_user_path(conn_or_endpoint, action, params)

project_user_path(conn_or_endpoint, action, id, params)

project_user_url(conn_or_endpoint, action)

project_user_url(conn_or_endpoint, action, params)

project_user_url(conn_or_endpoint, action, id, params)

role_path(conn_or_endpoint, action)

role_path(conn_or_endpoint, action, params)

role_path(conn_or_endpoint, action, id, params)

role_skill_path(conn_or_endpoint, action)

role_skill_path(conn_or_endpoint, action, params)

role_skill_path(conn_or_endpoint, action, id, params)

role_skill_url(conn_or_endpoint, action)

role_skill_url(conn_or_endpoint, action, params)

role_skill_url(conn_or_endpoint, action, id, params)

role_url(conn_or_endpoint, action)

role_url(conn_or_endpoint, action, params)

role_url(conn_or_endpoint, action, id, params)

skill_path(conn_or_endpoint, action)

skill_path(conn_or_endpoint, action, params)

skill_path(conn_or_endpoint, action, id, params)

skill_url(conn_or_endpoint, action)

skill_url(conn_or_endpoint, action, params)

skill_url(conn_or_endpoint, action, id, params)

slugged_route_path(conn_or_endpoint, action, slug)

slugged_route_path(conn_or_endpoint, action, slug, params)

slugged_route_url(conn_or_endpoint, action, slug)

slugged_route_url(conn_or_endpoint, action, slug, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

stripe_connect_account_path(conn_or_endpoint, action)

stripe_connect_account_path(conn_or_endpoint, action, id)

stripe_connect_account_path(conn_or_endpoint, action, id, params)

stripe_connect_account_url(conn_or_endpoint, action)

stripe_connect_account_url(conn_or_endpoint, action, id)

stripe_connect_account_url(conn_or_endpoint, action, id, params)

stripe_connect_events_path(conn_or_endpoint, action)

stripe_connect_events_path(conn_or_endpoint, action, params)

stripe_connect_events_url(conn_or_endpoint, action)

stripe_connect_events_url(conn_or_endpoint, action, params)

stripe_connect_plan_path(conn_or_endpoint, action)

stripe_connect_plan_path(conn_or_endpoint, action, id)

stripe_connect_plan_path(conn_or_endpoint, action, id, params)

stripe_connect_plan_url(conn_or_endpoint, action)

stripe_connect_plan_url(conn_or_endpoint, action, id)

stripe_connect_plan_url(conn_or_endpoint, action, id, params)

stripe_connect_subscription_path(conn_or_endpoint, action)

stripe_connect_subscription_path(conn_or_endpoint, action, id)

stripe_connect_subscription_path(conn_or_endpoint, action, id, params)

stripe_connect_subscription_url(conn_or_endpoint, action)

stripe_connect_subscription_url(conn_or_endpoint, action, id)

stripe_connect_subscription_url(conn_or_endpoint, action, id, params)

stripe_platform_card_path(conn_or_endpoint, action)

stripe_platform_card_path(conn_or_endpoint, action, id)

stripe_platform_card_path(conn_or_endpoint, action, id, params)

stripe_platform_card_url(conn_or_endpoint, action)

stripe_platform_card_url(conn_or_endpoint, action, id)

stripe_platform_card_url(conn_or_endpoint, action, id, params)

stripe_platform_customer_path(conn_or_endpoint, action)

stripe_platform_customer_path(conn_or_endpoint, action, id)

stripe_platform_customer_path(conn_or_endpoint, action, id, params)

stripe_platform_customer_url(conn_or_endpoint, action)

stripe_platform_customer_url(conn_or_endpoint, action, id)

stripe_platform_customer_url(conn_or_endpoint, action, id, params)

stripe_platform_events_path(conn_or_endpoint, action)

stripe_platform_events_path(conn_or_endpoint, action, params)

stripe_platform_events_url(conn_or_endpoint, action)

stripe_platform_events_url(conn_or_endpoint, action, params)

task_list_path(conn_or_endpoint, action)

task_list_path(conn_or_endpoint, action, params)

task_list_path(conn_or_endpoint, action, id, params)

task_list_task_path(conn_or_endpoint, action, task_list_id)

task_list_task_path(conn_or_endpoint, action, task_list_id, params)

task_list_task_path(conn_or_endpoint, action, task_list_id, number, params)

task_list_task_url(conn_or_endpoint, action, task_list_id)

task_list_task_url(conn_or_endpoint, action, task_list_id, params)

task_list_task_url(conn_or_endpoint, action, task_list_id, number, params)

task_list_url(conn_or_endpoint, action)

task_list_url(conn_or_endpoint, action, params)

task_list_url(conn_or_endpoint, action, id, params)

task_path(conn_or_endpoint, action)

task_path(conn_or_endpoint, action, params)

task_path(conn_or_endpoint, action, id, params)

task_skill_path(conn_or_endpoint, action)

task_skill_path(conn_or_endpoint, action, params)

task_skill_path(conn_or_endpoint, action, id, params)

task_skill_url(conn_or_endpoint, action)

task_skill_url(conn_or_endpoint, action, params)

task_skill_url(conn_or_endpoint, action, id, params)

task_url(conn_or_endpoint, action)

task_url(conn_or_endpoint, action, params)

task_url(conn_or_endpoint, action, id, params)

token_path(conn_or_endpoint, action)

token_path(conn_or_endpoint, action, params)

token_url(conn_or_endpoint, action)

token_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_category_path(conn_or_endpoint, action)

user_category_path(conn_or_endpoint, action, params)

user_category_path(conn_or_endpoint, action, id, params)

user_category_url(conn_or_endpoint, action)

user_category_url(conn_or_endpoint, action, params)

user_category_url(conn_or_endpoint, action, id, params)

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, id, params)

user_role_path(conn_or_endpoint, action)

user_role_path(conn_or_endpoint, action, params)

user_role_path(conn_or_endpoint, action, id, params)

user_role_url(conn_or_endpoint, action)

user_role_url(conn_or_endpoint, action, params)

user_role_url(conn_or_endpoint, action, id, params)

user_skill_path(conn_or_endpoint, action)

user_skill_path(conn_or_endpoint, action, params)

user_skill_path(conn_or_endpoint, action, id, params)

user_skill_url(conn_or_endpoint, action)

user_skill_url(conn_or_endpoint, action, params)

user_skill_url(conn_or_endpoint, action, id, params)

user_task_path(conn_or_endpoint, action)

user_task_path(conn_or_endpoint, action, params)

user_task_path(conn_or_endpoint, action, id, params)

user_task_url(conn_or_endpoint, action)

user_task_url(conn_or_endpoint, action, params)

user_task_url(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, id, params)