code_corps v0.0.1 CodeCorps.GitHub

Types

api_error_struct()

api_error_struct() ::
  CodeCorps.GitHub.APIError.t() |
  CodeCorps.GitHub.HTTPClientError.t()

body()

body() :: {:multipart, list()} | map()

headers()

headers() :: %{optional(String.t()) => String.t()} | %{}

method()

method() :: :get | :post | :put | :delete | :patch | :head

paginated_endpoint_error()

paginated_endpoint_error() ::
  Ecto.Changeset.t() |
  CodeCorps.GitHub.APIError.t() |
  CodeCorps.GitHub.HTTPClientError.t() |
  CodeCorps.GitHub.API.Errors.PaginationError.t()

Potential errors which can happen when retrieving data from a paginated endpoint.

If a new access token is required, then it is regenerated and stored into an installation, which can result in any of

Ecto.Changeset.t

CodeCorps.GitHub.APIError.t

CodeCorps.GitHub.HTTPClientError.t

Once that is done, the actual request is made, which can error out with

CodeCorps.GitHub.Errors.PaginationError.t

response()

response() :: {:ok, map()} | {:error, api_error_struct()}

Functions

get_all(endpoint, headers, options)

get_all(String.t(), headers(), list()) ::
  {:ok, [map()]} |
  {:error, CodeCorps.GitHub.API.Errors.PaginationError.t()} |
  {:error, api_error_struct()}

A low level utility function to make an authenticated request to a GitHub API endpoint which supports pagination, and fetch all the pages from that endpoint at once, by making parallel requests to each page and aggregating the results.

integration_request(method, endpoint, body, headers, options)

integration_request(method(), String.t(), body(), headers(), list()) :: response()

A low level utility function to make an authenticated request to the GitHub API on behalf of a GitHub App or integration

request(method, endpoint, body, headers, options)

request(method(), String.t(), body(), headers(), list()) :: response()

A low level utility function to make a direct request to the GitHub API.

user_access_token_request(code, state)

user_access_token_request(String.t(), String.t()) :: response()

A low level utility function to fetch a GitHub user’s OAuth access token