code_corps v0.0.1 CodeCorps.GitHub.Adapters.User

Used to adapt a GitHub issue payload into attributes for creating or updating a CodeCorps.Task.

Functions

from_github_user(payload)

from_github_user(map) :: map

Converts a Github user payload into a map of attributes suitable for creating or updating a CodeCorps.User

Any nil values are removed here, since we do not want to, for example, delete an existing email just because the github payload doesn’t have that data.