code_corps v0.0.1 CodeCorps.GitHub.Event.Installation

In charge of handling a GitHub Webhook payload for the Installation event type https://developer.github.com/v3/activity/events/types/#installationevent

Types

outcome()

outcome ::
  {:ok, CodeCorps.GithubAppInstallation.t} |
  {:error, :not_yet_implemented} |
  {:error, :unexpected_action} |
  {:error, :unexpected_payload} |
  {:error, :validation_error_on_syncing_installation} |
  {:error, :multiple_unprocessed_installations_found} |
  {:error, :github_api_error_on_syncing_repos} |
  {:error, :validation_error_on_deleting_removed_repos} |
  {:error, :validation_error_on_syncing_existing_repos} |
  {:error, :validation_error_on_marking_installation_processed}

Functions

handle(payload)

handle(map) :: outcome

Handles the “Installation” GitHub Webhook event.

The event could be of subtype created or deleted. Only the created variant is handled at the moment.

The process of handling the “created” event subtype is as follows

try to match the sender with an existing CodeCorps.User

call specific matching module depending on the user being matched or not

CodeCorps.GitHub.Event.Installation.MatchedUser.handle/2

CodeCorps.GitHub.Event.Installation.UnmatchedUser.handle/1

sync installation repositories using a third modules

CodeCorps.GitHub.Event.Installation.Repos.process/1

If everything goes as expected, an :ok tuple will be returned, with a CodeCorps.GithubAppInstallation, marked as “processed”.

If a step in the process failes, an :error tuple will be returned, where the second element is an atom indicating which step of the process failed.