code_corps v0.0.1 CodeCorps.GitHub.Event.Installation.Repos

In charge of processing repos during an installation event

Functions

process(installation)

process(CodeCorps.GithubAppInstallation.t) :: {:ok | :error, CodeCorps.GithubAppInstallation.t}

Fetches a list of repositories for a GithubAppInstallation from the GitHub API and matches up GithubRepo records stored locally using that list as the master list.

process_async(installation)

process_async(CodeCorps.GithubAppInstallation.t) :: {:ok, CodeCorps.GithubAppInstallation.t, Task.t}

Marks a GithubAppInstallation as “processing” and immediately returns it.

In the background, it fires a task to asynchronously call &process/1