code_corps v0.0.1 CodeCorps.GitHub.Event.InstallationRepositories

In charge of dealing with “InstallationRepositories” GitHub Webhook events

Functions

handle(github_event, payload)

handle(CodeCorps.GithubEvent.t, map) :: outcome

Handles an “InstallationRepositories” GitHub Webhook event. The event could be of subtype “added” or “removed” and is handled differently based on that.

InstallationRepositories::added will

find GithubAppInstallation

marks event as errored if not found

marks event as errored if payload does not have keys it needs, meaning there’s something wrong with our code and we need to updated

find or create GithubRepo records affected

InstallationRepositories::removed will

find GithubAppInstallation

marks event as errored if not found

marks event as errored if payload does not have keys it needs, meaning there’s something wrong with our code and we need to updated

find GithubRepo records affected and delete them

if there is no GithubRepo to delete, it just skips it

ProjectGithubRepo are deleted automatically, since they are set to on_delete: :delete_all