code_corps v0.0.1 CodeCorps.GitHub.Event.InstallationRepositories

In charge of handling a GitHub Webhook payload for the InstallationRepositories event type.

https://developer.github.com/v3/activity/events/types/#installationrepositoriesevent

Types

outcome()

outcome ::
  {:ok, [CodeCorps.GithubRepo.t]} |
  {:error, :unmatched_installation} |
  {:error, :unexpected_action} |
  {:error, :unexpected_payload} |
  {:error, :validation_error_on_syncing_repos} |
  {:error, :unexpected_transaction_outcome}

Functions

handle(payload)

handle(map) :: outcome

Handles an “InstallationRepositories” GitHub Webhook event. The event could be of subtype “added” or “removed” and is handled differently based on that.

the process of handling the “added” subtype is as follows

try to match with CodeCorps.GithubAppInstallation record

sync affected CodeCorps.GithubRepo records (update, create)

the process of handling the “removed” subtype is as follows

try to match with a CodeCorps.GithubAppInstallation record

delete affected CodeCorps.GithubRepo records, respecting the rules

if the GitHub payload for a repo is not matched with a record in our database, just skip deleting it

if the deleted CodeCorps.GithubRepo record is associated with CodeCorps.ProjectGithubRepo records, they are deleted automatically, due to on_delete: :delete_all set at the database level.