code_corps v0.0.1 CodeCorps.GitHub.Event.Issues

In charge of handling a GitHub Webhook payload for the Issues event type

https://developer.github.com/v3/activity/events/types/#issuesevent

Types

outcome()

outcome ::
  {:ok, [CodeCorps.Task.t]} |
  {:error, :not_fully_implemented} |
  {:error, :unexpected_action} |
  {:error, :unexpected_payload} |
  {:error, :repository_not_found} |
  {:error, :validation_error_on_inserting_user} |
  {:error, :multiple_github_users_matched_same_cc_user} |
  {:error, :validation_error_on_syncing_tasks} |
  {:error, :unexpected_transaction_outcome}

Functions

handle(payload)

handle(map) :: outcome

Handles the “Issues” GitHub webhook

The process is as follows

validate the payload is structured as expected

validate the action is properly supported

match payload with affected CodeCorps.GithubRepo record using CodeCorps.GitHub.Event.Common.RepoFinder

match with a CodeCorps.User using CodeCorps.GitHub.Event.Issues.UserLinker

for each CodeCorps.ProjectGithubRepo belonging to matched repo

match and update, or create a CodeCorps.Task on the associated CodeCorps.Project

If the process runs all the way through, the function will return an :ok tuple with a list of affected (created or updated) tasks.

If it fails, it will instead return an :error tuple, where the second element is the atom indicating a reason.