code_corps v0.0.1 CodeCorps.GitHub.Sync

Syncs events received from the GitHub API and also syncs entire GitHub repositories.

Types

outcome()

outcome() ::
  {:ok, [CodeCorps.Comment.t()]} |
  {:ok, CodeCorps.GithubPullRequest.t()} |
  {:ok, [CodeCorps.Task.t()]} |
  {:error, :repo_not_found, %{}} |
  {:error, :fetching_issue} |
  {:error, :fetching_pull_request} |
  {:error, :multiple_issue_users_match} |
  {:error, :multiple_comment_users_match} |
  {:error, :validating_github_pull_request} |
  {:error, :validating_github_issue} |
  {:error, :validating_github_comment} |
  {:error, :validating_user} |
  {:error, :validating_task} |
  {:error, :validating_comment} |
  {:error, :unexpected_transaction_outcome}

Functions

issue_comment_event(payload)

issue_comment_event(map()) :: outcome()

Syncs a GitHub IssueComment event.

For the deleted action

Deletes the related comment records with GitHub.Sync.Comment

For any other action

Finds the CodeCorps.GithubRepo

If it’s a pull request, it fetches the pull request from the GitHub API and syncs it with GitHub.Sync.PullRequest

Syncs the issue portion of the event with GitHub.Sync.Issue

Syncs the comment portion of the event with GitHub.Sync.Comment

https://developer.github.com/v3/activity/events/types/#issuecommentevent

issue_event(payload)

issue_event(map()) :: outcome()

Syncs a GitHub Issues event.

Finds the CodeCorps.GithubRepo

Syncs the issue portion of the event with GitHub.Sync.Issue

https://developer.github.com/v3/activity/events/types/#issuesevent

pull_request_event(payload)

pull_request_event(map()) :: outcome()

Syncs a GitHub PullRequest event.

Finds the CodeCorps.GithubRepo

Fetches the issue from the GitHub API

Syncs the pull request portion of the event with GitHub.Sync.PullRequest

Syncs the issue portion of the event with GitHub.Sync.Issue, using the changes passed in from the issue fetching and the pull request syncing

https://developer.github.com/v3/activity/events/types/#pullrequestevent

sync_repo(repo)

sync_repo(CodeCorps.GithubRepo.t()) :: {:ok, CodeCorps.GithubRepo.t()}

Syncs a GithubRepo with Code Corps.

Fetches and syncs records from the GitHub API for a given repository, marking progress of the sync state along the way.

Fetches the pull requests from the API

Creates or updates GithubPullRequest records (and their related GithubUser records)

Fetches the issues from the API

Creates or updates GithubIssue records, and relates them to any related GithubPullRequest records created previously (along with any related GithubUser records)

Fetches the comments from the API

Creates or updates GithubComment records (and their related GithubUser records)

Creates or updates User records for the GithubUser records

Creates or updates Task records, and relates them to any related GithubIssue and User records created previously

Creates or updates Comment records, and relates them to any related GithubComment and User records created previously