code_corps v0.0.1 CodeCorps.GitHub.Sync.PullRequest

In charge of finding a pull request to link with a GithubPullRequest record when processing a GitHub Pull Request payload.

The only entry point is create_or_update_pull_request/2.

Functions

create_or_update_pull_request(payload, github_repo)

create_or_update_pull_request(map(), CodeCorps.GithubRepo.t()) :: linking_result()

Finds or creates a GithubPullRequest using the data in a GitHub PullRequest payload.

The process is as follows:

Search for the pull request in our database with the payload data.

If we return a single GithubPullRequest, then the GithubPullRequest should be updated.

If there are no matching GithubPullRequest records, then a GithubPullRequestshould be created.