code_corps v0.0.1 CodeCorps.GitHub.TestHelpers

Functions

load_endpoint_fixture(id)

load_endpoint_fixture(String.t) :: map

load_event_fixture(id)

load_event_fixture(String.t) :: map

with_mock_api(mock_module, list) (macro)

with_mock_api(module, [{:do, function}]) :: any

Allows setting a mock Github API module for usage in specific tests To use it, define a module containing the methods expected to be called, then pass in the block of code expected to call it into the macro:

defmodule MyApiModule do
  def some_function, do: "foo"
end
with_mock_api(MyApiModule) do
  execute_code_calling_api
end