code_corps v0.0.1 CodeCorps.Helpers.Policy

Holds helpers for extracting record relationships and determining roles for authorization policies.

Functions

admin_or_higher?(role)

admin_or_higher?(String.t) :: boolean

Determines if provided string is equal to one of ["admin", "owner"]

contributor_or_higher?(role)

contributor_or_higher?(String.t) :: boolean

Determines if provided string is equal to one of ["contributor", "admin", "owner"]

get_membership(arg1, user)

get_membership(nil | Ecto.Changeset.t | CodeCorps.Project.t | CodeCorps.Organization.t | CodeCorps.StripeConnectAccount.t, CodeCorps.User.t) ::
  nil |
  CodeCorps.OrganizationMembership.t

Retrieves the specified user’s membership record, from a CodeCorps.Project struct or an Ecto.Changeset, containing an organization_id field, or from a CodeCorps.Organization struct

Returns CodeCorps.OrganizationMembership

get_project(arg1)

get_project(struct | Ecto.Changeset.t | any) :: CodeCorps.Project.t

Retrieves a project record, from a model struct, or an Ecto.Changeset containing a project_id field

Returns CodeCorps.Project

get_role(changeset)

get_role(nil | CodeCorps.OrganizationMembership.t | Ecto.Changeset.t) ::
  String.t |
  nil

Retrieves the role field, from a CodeCorps.OrganizationMembership struct or an Ecto.Changeset

Returns :string

get_task(arg1)

get_task(Ecto.Changeset.t | CodeCorps.TaskSkill.t | CodeCorps.UserTask.t) :: CodeCorps.Task.t

Retrieves task from associated record

owner?(arg1)

owner?(String.t) :: boolean

Determines if provided string is equal to “owner”

task_authored_by?(task, user)

task_authored_by?(CodeCorps.Task.t, CodeCorps.User.t) :: boolean

Determines if the provided task was authored by the provided user