code_corps v0.0.1 CodeCorps.StripeService.StripeConnectSubscriptionService

Functions

find_or_create(attributes)

Finds or creates a new Stripe.Subscription record on Stripe API, as well as an associated local StripeConnectSubscription record

Possible return values

{:ok, %StripeConnectSubscription{}} - the created record.

{:error, %Ecto.Changeset{}} - the record was not created due to validation issues.

{:error, :project_not_found}

{:error, :project_not_ready} - the associated project does not meed the prerequisites for receiving donations.

{:error, :user_not_found}

{:error, :user_not_ready} - the associated user does not meet the prerequisits to donate.

{:error, %Stripe.APIErrorResponse{}} - there was a problem with the stripe request

Side effects

If the subscription is created or found, associated project totals will get updated

If the subscription is created or found, associated donation goal states will be updated

update_from_stripe(stripe_id, connect_customer_id)

Updates an existing StripeConnectSubscription record by retrieving a Stripe.Subscription record and using that data as update parameters

Possible return values

{:ok, %StripeConnectSubscription{}} - the updated record.

{:error, %Stripe.APIErrorResponse{}} - there was a problem with the stripe request

{:error, :not_found} - one of the associated records was not found