code_corps v0.0.1 CodeCorps.StripeService.WebhookProcessing.WebhookProcessor

Used to process a Stripe webhook request.

Functions

process_async(json, handler)

Used to process a Stripe webhook event.

Receives the event JSON as the first parameter.

Since a webhook can be a platform or a connect webhook, the function requires the handler module as the second parameter.

Returns

{:ok, pid} if the event will be handled

{:error, :ignored_by_environment} if the event was ignored due to environment mismatch

Note

Stripe events can have their livemode property set to true or false. A livemode true event should be handled by the production environment, while all other environments handle livemode false events.