code_corps v0.0.1 CodeCorps.User

This module defines a user of the Code Corps app.

Types

t()

t() :: %CodeCorps.User{default_color: term(), id: term(), roles: term(), biography: term(), github_avatar_url: term(), github_auth_token: term(), state_transition: term(), github_app_installation: term(), user_skills: term(), github_user_id: term(), password_confirmation: term(), last_name: term(), project_users: term(), categories: term(), stripe_connect_subscriptions: term(), skills: term(), user_roles: term(), stripe_connect_customers: term(), stripe_platform_customer: term(), admin: term(), updated_at: term(), inserted_at: term(), user_categories: term(), sign_up_context: term(), username: term(), github_user: term(), email: term(), slugged_route: term(), github_app_installations: term(), stripe_platform_card: term(), state: term(), github_email: term(), encrypted_password: term(), github_id: term(), github_username: term(), cloudinary_public_id: term(), __meta__: term(), twitter: term(), type: term(), password: term(), website: term(), first_name: term()}

Functions

apply_state_transition(changeset, map)

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

check_email_availability(email)

check_username_availability(username)

registration_changeset(struct, params)

Builds a changeset for registering the user.

reset_password_changeset(struct, params)

update_changeset(struct, params)