code_corps v0.0.1 CodeCorps.User

This module defines a user of the Code Corps app.

Types

t()

t() :: %CodeCorps.User{__meta__: term(), biography: term(), github_email: term(), user_roles: term(), id: term(), first_name: term(), user_skills: term(), github_auth_token: term(), inserted_at: term(), email: term(), github_id_was: term(), user_categories: term(), updated_at: term(), stripe_connect_subscriptions: term(), github_user: term(), stripe_connect_customers: term(), website: term(), admin: term(), categories: term(), github_app_installations: term(), default_color: term(), skills: term(), password_confirmation: term(), project_users: term(), github_id: term(), github_avatar_url: term(), github_app_installation: term(), sign_up_context: term(), username: term(), state: term(), roles: term(), password: term(), cloudinary_public_id: term(), last_name: term(), state_transition: term(), stripe_platform_card: term(), github_username: term(), twitter: term(), stripe_platform_customer: term(), slugged_route: term(), type: term(), github_user_id: term(), encrypted_password: term()}

Functions

apply_state_transition(changeset, map)

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

check_email_availability(email)

check_username_availability(username)

full_name(user)

Returns full name of the given user.

registration_changeset(struct, params)

Builds a changeset for registering the user.

reset_password_changeset(struct, params)

update_changeset(struct, params)