code_corps v0.0.1 CodeCorps.User

This module defines a user of the Code Corps app.

Types

t()

t() :: %CodeCorps.User{__meta__: term, encrypted_password: term, stripe_connect_subscriptions: term, username: term, website: term, state_transition: term, user_categories: term, twitter: term, github_username: term, email: term, categories: term, github_app_installations: term, stripe_connect_customers: term, user_skills: term, stripe_platform_card: term, inserted_at: term, password: term, github_avatar_url: term, slugged_route: term, github_id: term, stripe_platform_customer: term, biography: term, github_auth_token: term, updated_at: term, state: term, project_users: term, password_confirmation: term, skills: term, sign_up_context: term, github_app_installation: term, id: term, first_name: term, last_name: term, default_color: term, user_roles: term, cloudinary_public_id: term, admin: term, github_email: term, roles: term}

Functions

apply_state_transition(changeset, map)

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

check_email_availability(email)

check_username_availability(username)

registration_changeset(struct, params)

Builds a changeset for registering the user.

reset_password_changeset(struct, params)

update_changeset(struct, params)